+++ platform/core/src/DICOMSR/SCOORD3D/utils/getSequenceAsArray.js
@@ -1,0 +1,4 @@
+const getSequenceAsArray = sequence =>
+  Array.isArray(sequence) ? sequence : [sequence];
+
+export default getSequenceAsArray;