+++ platform/core/src/classes/StudyPrefetcher.js
@@ -1,6 +1,4 @@
 import cornerstone from 'cornerstone-core';
-import cornerstoneTools from 'cornerstone-tools';
-
 import getImageId from '../utils/getImageId.js';
 
 const noop = () => {};
@@ -13,6 +11,7 @@ export class StudyPrefetcher {
     requestType: 'prefetch',
     preventCache: false,
     prefetchDisplaySetsTimeout: 300,
+    maxNumPrefetchRequests: 100,
     includeActiveDisplaySet: false,
   };
 
@@ -94,7 +93,11 @@ export class StudyPrefetcher {
     }
 
     this.element = element;
-    this.enabledElement = cornerstone.getEnabledElement(element);
+    try {
+      this.enabledElement = cornerstone.getEnabledElement(element);
+    } catch {
+      throw new Error('Failed to find the enabled element');
+    }
 
     this.stopPrefetching();
     this.prefetchDisplaySets(displaySetInstanceUID);
@@ -114,7 +117,11 @@ export class StudyPrefetcher {
    * @param {number} timeout
    */
   prefetchDisplaySetsAsync(element, timeout) {
-    this.enabledElement = cornerstone.getEnabledElement(element);
+    try {
+      this.enabledElement = cornerstone.getEnabledElement(element);
+    } catch {
+      throw new Error('Failed to find the enabled element');
+    }
     timeout = timeout || this.options.prefetchDisplaySetsTimeout;
     clearTimeout(this.prefetchDisplaySetsHandler);
     this.prefetchDisplaySetsHandler = setTimeout(() => {
@@ -144,6 +151,11 @@ export class StudyPrefetcher {
   prefetchImageIds(imageIds) {
     const nonCachedImageIds = this.filterCachedImageIds(imageIds);
     const imageLoadPoolManager = cornerstone.imageLoadPoolManager;
+
+    imageLoadPoolManager.maxNumRequests = {
+      ...imageLoadPoolManager.maxNumRequests,
+      prefetch: this.options.maxNumPrefetchRequests,
+    };
 
     let requestFn;
     if (this.options.preventCache) {